requirements.txt
coverage==5.0.3
csuffixtree==0.3.6
dataclasses_serialization==1.2.4
Flask==1.1.1
gunicorn==20.0.4
pylint==2.4.4
pytest==5.3.5
pytest-cov==2.8.1
python-Levenshtein-wheels==0.13.1
